--- app/src/main/java/org/sopt/dosopttemplate/data/local/SharedPreferenceMain.kt
@@ -1,9 +1,0 @@
-package org.sopt.dosopttemplate.data.local
-
-import org.sopt.dosopttemplate.domain.entity.UserEntity
-
-interface SharedPreferenceMain {
-    var user:UserEntity
-    var checkLogin: Boolean
-    fun clearPref()
-}
