+++ app/src/main/java/org/sopt/dosopttemplate/data/local/repo/AuthRepoImpl.kt
@@ -1,27 +1,27 @@
 package org.sopt.dosopttemplate.data.local.repo
 
-import org.sopt.dosopttemplate.data.local.SharedPreference
+import org.sopt.dosopttemplate.data.local.SharedPreferenceImpl
 import org.sopt.dosopttemplate.domain.entity.UserEntity
 import org.sopt.dosopttemplate.domain.repo.AuthRepo
 import javax.inject.Inject
 
 class AuthRepoImpl @Inject constructor(
-    private val sharedPreference: SharedPreference
+    private val sharedPreferenceImpl: SharedPreferenceImpl
 ) : AuthRepo {
 
-    override fun getUser(): UserEntity = sharedPreference.user
+    override fun getUser(): UserEntity = sharedPreferenceImpl.user
 
     override fun saveUser(userEntity: UserEntity) {
-        sharedPreference.user = userEntity
+        sharedPreferenceImpl.user = userEntity
     }
 
-    override fun checkLogin() = sharedPreference.checkLogin
+    override fun checkLogin() = sharedPreferenceImpl.checkLogin
     override fun saveCheckLogin(checkLogin: Boolean) {
-        sharedPreference.checkLogin = checkLogin
+        sharedPreferenceImpl.checkLogin = checkLogin
     }
 
     override fun clear() {
-        sharedPreference.clearPref()
+        sharedPreferenceImpl.clearPref()
     }
 
 }
